RenAIssance_VLMOCR_Tushar_Tyagi/requirements.txt
transformers>=4.40
torch>=2.1
bitsandbytes>=0.43
accelerate>=0.30
qwen-vl-utils
jiwer>=3.0
Pillow>=10.0
pyspellchecker>=0.7.1